Разделение AggregatorService.java
