+++ stats/aggregator/src/main/java/ru/practicum/aggregator/AggregatorService.java
@@ -7,37 +7,26 @@ import lombok.RequiredArgsConstructor;
 import lombok.experimental.FieldDefaults;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.avro.specific.SpecificRecordBase;
-import org.apache.kafka.clients.consumer.ConsumerRecord;
-import org.apache.kafka.clients.consumer.ConsumerRecords;
 import org.apache.kafka.clients.consumer.KafkaConsumer;
 import org.apache.kafka.common.errors.WakeupException;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
-import ru.practicum.ewm.stats.avro.ActionTypeAvro;
 import ru.practicum.ewm.stats.avro.EventSimilarityAvro;
 import ru.practicum.ewm.stats.avro.UserActionAvro;
 import ru.practicum.producer.KafkaProducerService;
 
 import java.time.Duration;
-import java.time.Instant;
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.Comparator;
 import java.util.List;
-import java.util.Map;
-import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 
-@Slf4j
-@RequiredArgsConstructor
 @Service
+@RequiredArgsConstructor
 @FieldDefaults(level = AccessLevel.PRIVATE)
+@Slf4j
 public class AggregatorService {
-    static final double VIEW_WEIGHT = 0.4;
-    static final double LIKE_WEIGHT = 1.0;
-    static final double REGISTER_WEIGHT = 0.8;
 
+    final AggregationProcessor processor;
     final KafkaConsumer<String, SpecificRecordBase> consumer;
     final KafkaProducerService kafkaProducer;
 
@@ -53,20 +42,10 @@ public class AggregatorService {
     final ExecutorService executor = Executors.newSingleThreadExecutor();
     volatile boolean running = true;
 
-    // Карта: ID мероприятия → (ID пользователя → максимальный вес действия)
-    final Map<Long, Map<Long, Double>> eventUserWeights = new ConcurrentHashMap<>();
-    // Карта: ID мероприятия → сумма всех весов пользователей
-    final Map<Long, Double> eventTotalWeights = new ConcurrentHashMap<>();
-    // Карта: ID мероприятия A → (ID мероприятия B → сумма минимальных весов общих пользователей)
-    final Map<Long, Map<Long, Double>> minWeightsSums = new ConcurrentHashMap<>();
-    // Карта: ID пользователя → список ID мероприятий, с которыми он взаимодействовал
-    final Map<Long, List<Long>> userEvents = new ConcurrentHashMap<>();
-
-
     @PostConstruct
     public void start() {
         log.info("Подписка consumer на топик {}", inputTopic);
-        consumer.subscribe(Collections.singletonList(inputTopic));
+        consumer.subscribe(List.of(inputTopic));
         executor.submit(this::processMessages);
     }
 
@@ -74,286 +53,28 @@ public class AggregatorService {
     public void stop() {
         running = false;
         consumer.wakeup();
-        try {
-            log.info("Завершение работы: отправка оставшихся сообщений...");
-            kafkaProducer.flush();
-        } catch (Exception e) {
-            log.error("Ошибка при завершении продюсера", e);
-        }
+        kafkaProducer.flush();
         executor.shutdown();
-        log.info("Kafka consumer остановлен");
     }
 
     private void processMessages() {
         while (running) {
             try {
-                ConsumerRecords<String, SpecificRecordBase> records = consumer.poll(Duration.ofMillis(pollTimeout));
-                log.info("Получено {} записей для обработки", records.count());
-                for (ConsumerRecord<String, SpecificRecordBase> record : records) {
-                    log.info("Получена запись {}", record);
-                    SpecificRecordBase value = record.value();
-                    if (value instanceof UserActionAvro userAction) {
-                        log.info("Обработка действия: {}", userAction);
-                        try {
-                            List<EventSimilarityAvro> similarities = processUserAction(userAction);
-                            similarities.forEach(similarity -> {
-                                log.info("Отправка similarityEvent {} в топик {}", similarity, outputTopic);
-                                kafkaProducer.send(similarity, outputTopic);
-                            });
-                        } catch (Exception e) {
-                            log.error("Ошибка обработки действия: {}", userAction, e);
-                        }
-                    } else {
-                        log.warn("Получено сообщение неизвестного типа: {}", value.getClass().getSimpleName());
+                var records = consumer.poll(Duration.ofMillis(pollTimeout));
+                for (var record : records) {
+                    if (record.value() instanceof UserActionAvro userAction) {
+                        List<EventSimilarityAvro> result = processor.process(userAction);
+                        result.forEach(sim -> kafkaProducer.send(sim, outputTopic));
                     }
                 }
-                if (!records.isEmpty()) {
-                    consumer.commitSync();
-                }
+                if (!records.isEmpty()) consumer.commitSync();
                 kafkaProducer.flush();
             } catch (WakeupException e) {
                 if (!running) break;
             } catch (Exception e) {
-                log.error("Ошибка в основном цикле обработки", e);
+                log.error("Ошибка в цикле обработки", e);
             }
         }
-    }
-
-    /**
-     * Этот метод определяет тип действия и делегирует обработку соответствующему методу:
-     *
-     * Если мероприятие новое - вызывает handleNewEvent
-     *
-     * Если пользователь впервые взаимодействует с существующим мероприятием - handleNewUserForEvent
-     *
-     * Если это обновление веса существующего действия - handleWeightUpdate
-     */
-    private List<EventSimilarityAvro> processUserAction(UserActionAvro userAction) {
-        long userId = userAction.getUserId();
-        long eventId = userAction.getEventId();
-        double newWeight = getActionWeight(userAction.getActionType());
-
-        if (!eventUserWeights.containsKey(eventId)) {
-            // Первое взаимодействие с этим мероприятием
-            return handleNewEvent(userId, eventId, newWeight);
-        } else {
-            Map<Long, Double> eventUsers = eventUserWeights.get(eventId);
-            if (!eventUsers.containsKey(userId)) {
-                // Первое взаимодействие пользователя с этим мероприятием
-                return handleNewUserForEvent(userId, eventId, newWeight);
-            } else {
-                double oldWeight = eventUsers.get(userId);
-                if (newWeight > oldWeight) {
-                    // Обновление веса на более значимое действие
-                    return handleWeightUpdate(userId, eventId, oldWeight, newWeight);
-                }
-            }
-        }
-        return Collections.emptyList();
-    }
-
-    /**
-     *
-     * @param userId
-     * @param eventId
-     * @param newWeight
-     * @return
-     * Создает новую запись для мероприятия в eventUserWeights
-     *
-     * Инициализирует общую сумму весов для мероприятия
-     *
-     * Добавляет мероприятие в список мероприятий пользователя
-     *
-     * Обновляет суммы минимумов для всех пар с новым мероприятием
-     *
-     * Рассчитывает и возвращает новые значения сходства
-     */
-    private List<EventSimilarityAvro> handleNewEvent(long userId, long eventId, double newWeight) {
-        // Создаем новую запись для мероприятия
-        Map<Long, Double> newMap = new ConcurrentHashMap<>();
-        newMap.put(userId, newWeight);
-        eventUserWeights.put(eventId, newMap);
-
-        // Обновляем общую сумму весов
-        eventTotalWeights.put(eventId, newWeight);
-
-        // Добавляем мероприятие в историю пользователя
-        userEvents.compute(userId, (k, events) -> {
-            if (events == null) {
-                events = new ArrayList<>();
-            }
-            events.add(eventId);
-            return events;
-        });
-
-        // Обновляем суммы минимумов для всех мероприятий пользователя
-        updateMinWeightsForUser(userId, eventId, 0.0, newWeight);
-
-        // Генерируем события сходства для всех пар с новым мероприятием
-        return generateSimilaritiesForUserEvents(userId, eventId);
-    }
-    // Первое взаимодействие пользователя с этим мероприятием
-    private List<EventSimilarityAvro> handleNewUserForEvent(long userId, long eventId, double newWeight) {
-        // Добавляем пользователя к существующему мероприятию
-        eventUserWeights.get(eventId).put(userId, newWeight);
-
-        // Обновляем общую сумму весов
-        eventTotalWeights.merge(eventId, newWeight, Double::sum);
-
-        // Добавляем мероприятие в историю пользователя
-        userEvents.compute(userId, (k, events) -> {
-            if (events == null) {
-                events = new ArrayList<>();
-            }
-            events.add(eventId);
-            return events;
-        });
-
-        // Обновляем суммы минимумов для всех мероприятий пользователя
-        updateMinWeightsForUser(userId, eventId, 0.0, newWeight);
-
-        // Генерируем события сходства для всех пар с новым мероприятием
-        return generateSimilaritiesForUserEvents(userId, eventId);
-    }
-
-    private List<EventSimilarityAvro> handleWeightUpdate(long userId, long eventId, double oldWeight, double newWeight) {
-        // Обновляем вес пользователя для мероприятия
-        eventUserWeights.get(eventId).put(userId, newWeight);
-
-        // Обновляем общую сумму весов
-        double delta = newWeight - oldWeight;
-        eventTotalWeights.merge(eventId, delta, Double::sum);
-
-        // Обновляем суммы минимумов для всех мероприятий пользователя
-        updateMinWeightsForUser(userId, eventId, oldWeight, newWeight);
-
-        // Генерируем события сходства для всех пар с обновленным мероприятием
-        return generateSimilaritiesForUserEvents(userId, eventId);
-    }
-//суммы минимумов
-
-    /**
-     *
-     * @param userId
-     * @param eventId
-     * @param oldWeight
-     * @param newWeight
-     * Когда вес пользователя для мероприятия изменяется, это может повлиять на минимальные веса во всех парах, где участвует это мероприятие и данный пользователь.
-     *
-     * Например, если у пользователя было:
-     *
-     * Вес для мероприятия A: 0.5 (старый)
-     *
-     * Вес для мероприятия B: 0.7
-     *
-     * Минимум был min(0.5, 0.7) = 0.5
-     *
-     * Если вес для A изменился на 0.8:
-     *
-     * Новый минимум min(0.8, 0.7) = 0.7
-     *
-     * Разница: 0.7 - 0.5 = 0.2
-     *
-     * Значит, общая сумма минимумов для пары (A,B) должна увеличиться на 0.2
-     */
-    private void updateMinWeightsForUser(long userId, long eventId, double oldWeight, double newWeight) {
-        // Получаем список всех мероприятий, с которыми взаимодействовал пользователь
-        List<Long> userEventIds = userEvents.get(userId);
-        if (userEventIds == null) return;
-
-        for (Long otherEventId : userEventIds) {
-            // Пропускаем текущее мероприятие (eventId) - смотрим только пары с другими
-            if (otherEventId.equals(eventId)) continue;
-            // Получаем вес пользователя для другого мероприятия в паре
-            double otherWeight = eventUserWeights.get(otherEventId).get(userId);
-            // Вычисляем предыдущий минимум для этой пары весов
-            double oldMin = Math.min(oldWeight, otherWeight);
-            // Вычисляем новый минимум для этой пары весов
-            double newMin = Math.min(newWeight, otherWeight);
-            double delta = newMin - oldMin;
-
-            if (delta != 0) {
-                updateMinWeightsSum(eventId, otherEventId, delta);
-            }
-        }
-    }
-
-    private List<EventSimilarityAvro> generateSimilaritiesForUserEvents(long userId, long eventId) {
-        List<EventSimilarityAvro> similarities = new ArrayList<>();
-        List<Long> userEventIds = userEvents.get(userId);
-        if (userEventIds == null) return similarities;
-
-        for (Long otherEventId : userEventIds) {
-            if (otherEventId.equals(eventId)) continue;
-
-            double numerator = getMinWeightsSum(eventId, otherEventId);
-            double denominator = Math.sqrt(eventTotalWeights.get(eventId)) * Math.sqrt(eventTotalWeights.get(otherEventId));
-
-            if (denominator == 0) {
-                continue;
-            }
-
-            double similarity = numerator / denominator;
-
-            long firstEventId = Math.min(eventId, otherEventId);
-            long secondEventId = Math.max(eventId, otherEventId);
-
-            similarities.add(EventSimilarityAvro.newBuilder()
-                    .setEventA(firstEventId)
-                    .setEventB(secondEventId)
-                    .setScore(similarity)
-                    .setTimestamp(Instant.now())
-                    .build());
-        }
-
-        similarities.sort(Comparator
-                .comparing(EventSimilarityAvro::getEventA)
-                .thenComparing(EventSimilarityAvro::getEventB));
-
-        return similarities;
-    }
-
-    /**
-     *
-     * @param eventA
-     * @param eventB
-     * @param delta
-     * Этот метод обновляет сумму минимальных весов (S_min) для пары мероприятий в специальной структуре данных, гарантируя при этом:
-     *
-     * Упорядоченное хранение пар мероприятий (чтобы избежать дублирования)
-     *
-     * обновление значений в concurrent
-     *
-     * Эффективное изменение суммы минимумов на заданную дельту
-     */
-    private void updateMinWeightsSum(long eventA, long eventB, double delta) {
-        long first = Math.min(eventA, eventB);
-        long second = Math.max(eventA, eventB);
-
-        minWeightsSums.compute(first, (k, innerMap) -> {
-            // Если для первого мероприятия еще нет внутренней карты - создаем
-            if (innerMap == null) innerMap = new ConcurrentHashMap<>();
-            // Обновляем значение для второго мероприятия:
-            // - Если записи не было - создаем с значением delta
-            // - Если запись была - добавляем delta к существующему значению
-            innerMap.merge(second, delta, Double::sum);
-            return innerMap;
-        });
-    }
-
-    private double getMinWeightsSum(long eventA, long eventB) {
-        long first = Math.min(eventA, eventB);
-        long second = Math.max(eventA, eventB);
-
-        return minWeightsSums.getOrDefault(first, Collections.emptyMap())
-                .getOrDefault(second, 0.0);
-    }
-
-    private double getActionWeight(ActionTypeAvro actionType) {
-        return switch (actionType) {
-            case VIEW -> VIEW_WEIGHT;
-            case LIKE -> LIKE_WEIGHT;
-            case REGISTER -> REGISTER_WEIGHT;
-        };
     }
 }
+
